updated commands.rs to perform command decoding.
The command decoder has been implemented. It can
decode the oscilloscope_waveform draw command, character
draw command, rectangle draw command and the retrieval
of system info from the serial input. More error checking
needs to be implemented as well as testing it in a real
scenario, but as of right now, it works.