added display.rs and added plugins.
Each section of the pipeline now has it's own plugin. The
lib.rs file exposes a SystemSet which will be used to define
the ordering of systems to restrict the writing of data to
resources and reading from resources in each section from
systems that act in each section.

Also, serialport.rs has now been renamed to serial.rs.